Working new Chinese Remainder + profiled multiprocess simulations

+++ python-simulation/TaskGenerator.py
@@ -61,7 +61,7 @@ def generateTasksFrom(utilizations, T_LCM, Tmin, Tmax, synchronous, constrDeadli
 		O = 0 if synchronous else max(0, int(random.normalvariate(Tmin, (Tmax-Tmin)/2)))
 		T = random.randrange(Tmin, Tmax) if T_LCM == -1 else random.choice(divisors)
 		C = max(1, int(round(math.floor(u*T))))
-		D = random.randint(T-(T-C)/constrDeadlineFactor, T)
+		D = random.randint(int(T-(T-C)/constrDeadlineFactor), T)
 		tasks.append(Task.Task(O, C, D, T))
 
 	# translate offset so that Omin=0
